[skip netlify] Extract non-component exports to utility and type files
