+++ app/components/AmenityReportModal.tsx
@@ -3,7 +3,7 @@
 import { useState } from 'react'
 
 import { Checkbox, Dialog, DialogPanel, DialogTitle } from '@headlessui/react'
-import { amenityMap } from './AmenityChip'
+import { amenityMap } from '@/app/data/amenities'
 
 interface Props {
   isOpen: boolean
@@ -50,7 +50,7 @@ export default function AmenityReportModal({ isOpen, onClose, onSuccess, ameniti
             className="absolute top-4 right-4 text-stone-400 hover:text-stone-600 transition-colors"
             aria-label="Close"
           >
-            <svg xmlns="http://www.w3.org/2000/svg" className="h-5 w-5" viewBox="0 0 20 20" fill="currentColor">
+            <svg xmlns="http://www.w3.org/2000/svg" className="size-5" viewBox="0 0 20 20" fill="currentColor">
               <path
                 fillRule="evenodd"
                 d="M4.293 4.293a1 1 0 011.414 0L10 8.586l4.293-4.293a1 1 0 111.414 1.414L11.414 10l4.293 4.293a1 1 0 01-1.414 1.414L10 11.414l-4.293 4.293a1 1 0 01-1.414-1.414L8.586 10 4.293 5.707a1 1 0 010-1.414z"
