+++ app/components/DirectionsButton.tsx
@@ -1,11 +1,9 @@
 import { MapPin } from 'lucide-react'
+import { getGoogleMapsUrl } from '@/app/utils/maps'
 
 interface DirectionsButtonProps {
   coordinates: [number, number]
 }
-
-export const getGoogleMapsUrl = (coordinates: { latitude: number; longitude: number }) =>
-  `https://www.google.com/maps?q=${coordinates.longitude},${coordinates.latitude}`
 
 export default function DirectionsButton({ coordinates }: DirectionsButtonProps) {
   return (
@@ -18,7 +16,7 @@ export default function DirectionsButton({ coordinates }: DirectionsButtonProps)
       rel="noopener noreferrer"
       className="inline-flex items-center gap-1.5 bg-white hover:bg-stone-50 text-stone-800 px-4 py-2.5 rounded-3xl text-sm font-medium border border-stone-200 transition-colors"
     >
-      <MapPin className="h-4 w-4" />
+      <MapPin className="size-4" />
       Directions
     </a>
   )
