+++ app/utils/distance.ts
@@ -1,0 +1,11 @@
+export const DISTANCE_UNITS = { Meters: 'Meters', Miles: 'Miles' }
+
+export const roundDistance = ({ units, distance }: { units: string; distance: number }) => {
+  if (units === 'Miles') return Math.round(distance * 100) / 100
+  if (units === 'Meters') return Math.round(distance)
+}
+
+export const generateDistanceText = ({ units, distance }: { units: string; distance: string }) => {
+  const parsedDistance = parseFloat(distance)
+  return `${roundDistance({ units, distance: parsedDistance })} ${units.toLowerCase()} away`
+}
