+++ app/utils/maps.ts
@@ -1,0 +1,2 @@
+export const getGoogleMapsUrl = (coordinates: { latitude: number; longitude: number }) =>
+  `https://www.google.com/maps?q=${coordinates.longitude},${coordinates.latitude}`
